+++ packages/plugin-inquirer-cli/test/Core.mocks.js
@@ -30,10 +30,9 @@ class CoreMock {
       stop: this._sandbox.stub().resolves(),
       restartServer: this._sandbox.stub().resolves(),
       onChangeAlerts: this._sandbox.stub().returns(doNothing),
-      onChangeMocks: this._sandbox.stub().returns(doNothing),
       addRouter: this._sandbox.stub(),
       addSetting: this._sandbox.stub(),
-      mocks: {
+      mock: {
         current: "foo-current",
         plainMocks: [],
         plainRoutes: [],
@@ -42,6 +41,10 @@ class CoreMock {
         ids: [],
         useRouteVariant: this._sandbox.stub(),
         restoreRouteVariants: this._sandbox.stub(),
+        onChange: this._sandbox.stub().returns(doNothing),
+      },
+      server: {
+        restart: this._sandbox.stub(),
       },
       config: new ConfigMock().stubs.instance,
       alerts: {