+++ test/core-e2e/src/fixtures/config-file-with-plugins/TraceMocksPlugin.js
@@ -3,7 +3,7 @@ class Plugin {
     return "trace-mocks";
   }
 
-  constructor({ mocks, config, logger }) {
+  constructor({ mock, config, logger }) {
     this._traceMocks = config.addOption({
       name: "traceMocks",
       type: "boolean",
@@ -12,8 +12,8 @@ class Plugin {
     });
     this._traceMocks.onChange(this._onChangeTraceMocks.bind(this));
 
-    this._mocks = mocks;
-    this._onChangeMocks = this._onChangeMocks.bind(this);
+    this._mock = mock;
+    this._onChangeMock = this._onChangeMock.bind(this);
     this._logger = logger;
   }
 
@@ -21,15 +21,15 @@ class Plugin {
     return "trace-mocks";
   }
 
-  init({ onChangeMocks, logger }) {
+  init({ mock, logger }) {
     this._enabled = this._traceMocks.value;
-    this._removeChangeMocksListener = onChangeMocks(this._onChangeMocks);
+    this._removeChangeMockListener = mock.onChange(this._onChangeMock);
     logger.debug(`traceMocks initial value is ${this._traceMocks.value}`);
   }
 
   traceMocks() {
     if (this._enabled && this._started) {
-      this._logger.info(`There are ${this._mocks.plainMocks.length} mocks available`);
+      this._logger.info(`There are ${this._mock.plainMocks.length} mocks available`);
     }
   }
 
@@ -48,7 +48,7 @@ class Plugin {
     this._enabled = this._traceMocks.value;
   }
 
-  _onChangeMocks() {
+  _onChangeMock() {
     this.traceMocks();
   }
 }